requirements.txt
yfinance
alpha_vantage
pandas
numpy
stocktrends
MetaTrader5
statsmodels
nltk
vaderSentiment
textblob
scikit-learn
alpaca-py